requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
langchain==0.2.16
langchain-core==0.2.38
langchain-google-genai==2.1.10
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
